backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
openai==1.3.7
whisper==1.1.10
pytesseract==0.3.10
Pillow==10.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.21.0
pandas>=2.0.0
httpx==0.25.2
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
